requirements.txt
biopython>=1.81
numpy
matplotlib
pandas
